Implement Brou api provider access

+++ src/ExchangeRate/Providers/Models/CurrencyPairRate.cs
@@ -3,4 +3,4 @@ namespace ExchangeRate.Providers.Models;
 /// <summary>
 ///     Represents a currency pair with its corresponding exchange rate.
 /// </summary>
-public record CurrencyPairRate(string FromCurrency, string ToCurrency, double Rate, DateTime UpdatedAt);
+public record CurrencyPairRate(string FromCurrency, string ToCurrency, decimal Rate, DateTime UpdatedAt);
